# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Post:
    properties:
      author_id:
        type: integer
      content:
        type: string
      title:
        type: string
    type: object
  models.User:
    properties:
      email:
        type: string
      nickname:
        type: string
      password:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: loeken@internetz.me
    name: loeken
  description: This is a sample service for a blog
  termsOfService: loekenIsKing
  title: Rest Blog
  version: "1.0"
paths:
  /auth:
    post:
      consumes:
      - application/json
      description: Allows login to the application by generating a JWT
      parameters:
      - description: Login
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.User'
      responses:
        "200":
          description: Token
          schema:
            type: string
      summary: Login to the application, generates a JWT
      tags:
      - auth
  /post:
    get:
      consumes:
      - application/json
      description: lists all posts
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Post'
            type: array
      summary: lists all posts
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: Creates a new post
      parameters:
      - description: Json body containing post data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Post'
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Post'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Creates a new post
      tags:
      - posts
  /post/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: 'Format: ''Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhdXRob3JpemVkIjp0cnVlLCJleHAiOjE2MDQ1MzExMDIsInVzZXJfaWQiOjF9.O63ZS_Poy29dDdcZqHDN0XeMPbYPX-Vfyl_FPfsMTvQ'''
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: Entity Count
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Deletes a post
      tags:
      - posts
    get:
      consumes:
      - application/json
      description: list post by id
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
      summary: list post by id
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: Updates a post
      parameters:
      - description: Json body containing post data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Post'
      - description: 'Format: ''Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhdXRob3JpemVkIjp0cnVlLCJleHAiOjE2MDQ1MzExMDIsInVzZXJfaWQiOjF9.O63ZS_Poy29dDdcZqHDN0XeMPbYPX-Vfyl_FPfsMTvQ'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
      security:
      - ApiKeyAuth: []
      summary: Updates a post
      tags:
      - posts
  /user:
    get:
      consumes:
      - application/json
      description: lists all users
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      summary: lists all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Creates a new User account
      parameters:
      - description: Json body containing user credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.User'
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      summary: Creates a new User account
      tags:
      - users
  /user/{id}:
    delete:
      consumes:
      - application/json
      description: delete specific user by id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: 'Format: ''Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhdXRob3JpemVkIjp0cnVlLCJleHAiOjE2MDQ1MzExMDIsInVzZXJfaWQiOjF9.O63ZS_Poy29dDdcZqHDN0XeMPbYPX-Vfyl_FPfsMTvQ'''
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/models.User'
      security:
      - ApiKeyAuth: []
      summary: delete specific user by id
      tags:
      - users
    get:
      consumes:
      - application/json
      description: list specific user by id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      summary: list specific user by id
      tags:
      - users
    put:
      consumes:
      - application/json
      description: update specific user by id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Json body containing user credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.User'
      - description: 'Format: ''Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhdXRob3JpemVkIjp0cnVlLCJleHAiOjE2MDQ1MzExMDIsInVzZXJfaWQiOjF9.O63ZS_Poy29dDdcZqHDN0XeMPbYPX-Vfyl_FPfsMTvQ'''
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      security:
      - ApiKeyAuth: []
      summary: update specific user by id
      tags:
      - users
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"